+++ packages/cli/src/commands/project/create-project.ts
@@ -45,7 +45,8 @@ export default class Create_project extends Command {
       ROOT_API_URL_PROD
     ).catch((e) => this.error(e));
 
+    const [account, name] = result.key.split('/');
     this.log(`Successfully created project: ${result.key}
-    \nProject Url: ${BASE_PROJECT_URL}/project/${result.key}`);
+    \nProject Url: ${BASE_PROJECT_URL}/orgs/${account}/projects/${name}/deployments`);
   }
 }