+++ packages/cli/src/controller/project-controller.ts
@@ -7,6 +7,12 @@ import {errorHandle} from '../utils';
 interface createProjectType {
   key: string;
 }
+export const suffixFormat = (value: string) => {
+  return value
+    .replace(/(^\s*)|(\s*$)/g, '')
+    .replace(/\s+/g, '-')
+    .toLowerCase();
+};
 export async function createProject(
   organization: string,
   subtitle: string,
@@ -31,7 +37,7 @@ export async function createProject(
           apiVersion: `v${apiVersion}`,
           description: description,
           gitRepository: gitRepository,
-          key: `${organization}/${project_name}`,
+          key: `${organization}/${suffixFormat(project_name)}`,
           logoUrl: logoUrl,
           name: project_name,
           subtitle: subtitle,