+++ packages/cli/src/controller/publish-controller.ts
@@ -79,7 +79,7 @@ async function replaceFileReferences<T>(
     return (await Promise.all(
       input.map((val) => replaceFileReferences(projectDir, val, authToken, ipfs))
     )) as unknown as T;
-  } else if (typeof input === 'object') {
+  } else if (typeof input === 'object' && input !== null) {
     if (input instanceof Map) {
       input = mapToObject(input) as T;
     }
@@ -115,7 +115,7 @@ export async function uploadFile(
   }
   let ipfsClusterCid: string;
   try {
-    ipfsClusterCid = await UploadFileByCluster(
+    ipfsClusterCid = await uploadFileByCluster(
       determineStringOrFsStream(content) ? await fs.promises.readFile(content.path, 'utf8') : content,
       authToken
     );
@@ -134,7 +134,7 @@ function determineStringOrFsStream(toBeDetermined: unknown): toBeDetermined is f
   return !!(toBeDetermined as fs.ReadStream).path;
 }
 
-async function UploadFileByCluster(content: string, authToken: string): Promise<string> {
+async function uploadFileByCluster(content: string, authToken: string): Promise<string> {
   const bodyFormData = new FormData();
   bodyFormData.append('data', content);
   const result = (
@@ -151,7 +151,17 @@ async function UploadFileByCluster(content: string, authToken: string): Promise<
       maxContentLength: 50 * 1024 * 1024,
     })
   ).data as ClusterResponseData;
-  return result.cid?.['/'];
+
+  if (typeof result.cid === 'string') {
+    return result.cid;
+  }
+  const cid = result.cid?.['/'];
+
+  if (!cid) {
+    throw new Error('Failed to get CID from response');
+  }
+
+  return cid;
 }
 
 function mapToObject(map: Map<string | number, unknown>): Record<string | number, unknown> {
@@ -165,15 +175,15 @@ function mapToObject(map: Map<string | number, unknown>): Record<string | number
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 function isFileReference(value: any): value is FileReference {
-  return value.file && typeof value.file === 'string';
+  return value?.file && typeof value.file === 'string';
 }
 
 interface ClusterResponseData {
   name: string;
-  cid: cidSpec;
+  cid: CidSpec | string;
   size: number;
 }
 // cluster response cid stored as {'/': 'QmVq2bqunmkmEmMCY3x9U9kDcgoRBGRbuBm5j7XKZDvSYt'}
-interface cidSpec {
+interface CidSpec {
   '/': string;
 }