+++ packages/node-core/src/db/db.module.ts
@@ -3,10 +3,9 @@
 
 import {DynamicModule, Global} from '@nestjs/common';
 import {Sequelize, Options as SequelizeOption} from 'sequelize';
-import * as entities from '../entities';
+import {NodeConfig} from '../configure/NodeConfig';
 import {getLogger} from '../logger';
 import {delay} from '../utils/promise';
-import {getYargsOption} from '../yargs';
 
 export interface DbOption {
   host: string;
@@ -17,6 +16,14 @@ export interface DbOption {
 }
 
 const logger = getLogger('db');
+
+const DEFAULT_DB_OPTION: DbOption = {
+  host: process.env.DB_HOST ?? '127.0.0.1',
+  port: process.env.DB_PORT ? Number(process.env.DB_PORT) : 5432,
+  username: process.env.DB_USER ?? 'postgres',
+  password: process.env.DB_PASS ?? 'postgres',
+  database: process.env.DB_DATABASE ?? 'postgres',
+};
 
 async function establishConnection(sequelize: Sequelize, numRetries: number): Promise<void> {
   try {
@@ -32,37 +39,65 @@ async function establishConnection(sequelize: Sequelize, numRetries: number): Pr
   }
 }
 
-const sequelizeFactory = (option: SequelizeOption) => async () => {
+const sequelizeFactory = (option: SequelizeOption, migrate: boolean) => async () => {
   const sequelize = new Sequelize(option);
   const numRetries = 5;
   await establishConnection(sequelize, numRetries);
-  for (const factoryFn of Object.keys(entities).filter((k) => /Factory$/.exec(k))) {
-    entities[factoryFn as keyof typeof entities](sequelize);
-  }
-  const {migrate} = getYargsOption().argv;
   await sequelize.sync({alter: migrate});
   return sequelize;
 };
 
 @Global()
 export class DbModule {
-  static forRoot(option: DbOption): DynamicModule {
-    const {argv} = getYargsOption();
+  static forRootWithConfig(nodeConfig: NodeConfig, option: DbOption = DEFAULT_DB_OPTION): DynamicModule {
+    const logger = getLogger('db');
+
+    const factory = sequelizeFactory(
+      {
+        ...option,
+        dialect: 'postgres',
+        logging: nodeConfig.debug
+          ? (sql: string, timing?: number) => {
+              logger.debug(sql);
+            }
+          : false,
+      },
+      nodeConfig.migrate
+    )();
+
+    return {
+      module: DbModule,
+      providers: [
+        {
+          provide: Sequelize,
+          useFactory: () => factory,
+        },
+      ],
+      exports: [Sequelize],
+    };
+  }
+
+  static forRoot(option: DbOption = DEFAULT_DB_OPTION): DynamicModule {
     const logger = getLogger('db');
     return {
       module: DbModule,
       providers: [
         {
           provide: Sequelize,
-          useFactory: sequelizeFactory({
-            ...option,
-            dialect: 'postgres',
-            logging: argv.debug
-              ? (sql: string, timing?: number) => {
-                  logger.debug(sql);
-                }
-              : false,
-          }),
+          useFactory: (nodeConfig: NodeConfig) =>
+            sequelizeFactory(
+              {
+                ...option,
+                dialect: 'postgres',
+                logging: nodeConfig.debug
+                  ? (sql: string, timing?: number) => {
+                      logger.debug(sql);
+                    }
+                  : false,
+              },
+              nodeConfig.migrate
+            )(),
+          inject: [NodeConfig],
         },
       ],
       exports: [Sequelize],