+++ packages/node-core/src/index.ts
@@ -3,11 +3,9 @@
 
 export * from './api.service';
 export * from './logger';
-export * from './yargs';
 export * from './profiler';
 export * from './events';
 export * from './configure';
-export * from './entities';
 export * from './db';
 export * from './meta';
 export * from './utils';