+++ packages/node-core/src/logger.ts
@@ -1,21 +1,25 @@
 // Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { isMainThread, threadId } from 'node:worker_threads';
-import { LoggerService } from '@nestjs/common';
-import { Logger } from '@subql/utils';
+import {isMainThread, threadId} from 'node:worker_threads';
+import {LoggerService} from '@nestjs/common';
+import {Logger} from '@subql/utils';
 import Pino from 'pino';
-import { argv } from './yargs';
 
-const outputFmt = argv('output-fmt') as 'json' | 'colored';
-const debug = argv('debug');
-const logLevel = argv('log-level') as string | undefined;
+let logger: Logger;
 
-const logger = new Logger({
-  level: debug ? 'debug' : logLevel,
-  outputFormat: outputFmt,
-  nestedKey: 'payload',
-});
+export function initLogger(debug = false, outputFmt?: 'json' | 'colored', logLevel?: string): void {
+  logger = new Logger({
+    level: debug ? 'debug' : logLevel,
+    outputFormat: outputFmt,
+    nestedKey: 'payload',
+  });
+}
+
+// Init logger for tests
+if ((global as any).__TEST__) {
+  initLogger();
+}
 
 export function getLogger(category: string): Pino.Logger {
   return logger.getLogger(category);
@@ -26,13 +30,11 @@ export function setLevel(level: Pino.LevelWithSilent): void {
 }
 
 export class NestLogger implements LoggerService {
-  private logger = logger.getLogger(
-    `nestjs${isMainThread ? '-0' : `-#${threadId}`}`,
-  );
+  private logger = logger.getLogger(`nestjs${isMainThread ? '-0' : `-#${threadId}`}`);
 
   error(message: any, trace?: string): void {
     if (trace) {
-      this.logger.error({ trace }, message);
+      this.logger.error({trace}, message);
     } else {
       this.logger.error(message);
     }