+++ packages/node/src/indexer/indexer.manager.spec.ts
@@ -10,7 +10,6 @@ import {
   StoreService,
   PoiService,
   MmrService,
-  SubqueryFactory,
   NodeConfig,
 } from '@subql/node-core';
 import { GraphQLSchema } from 'graphql';
@@ -124,17 +123,18 @@ function testSubqueryProject_2(): SubqueryProject {
   };
 }
 
-function createIndexerManager(project: SubqueryProject): IndexerManager {
+function createIndexerManager(
+  project: SubqueryProject,
+  nodeConfig: NodeConfig,
+): IndexerManager {
   const sequilize = new Sequelize();
   const eventEmitter = new EventEmitter2();
-
   const apiService = new ApiService(project, eventEmitter);
-  const dsProcessorService = new DsProcessorService(project);
+  const dsProcessorService = new DsProcessorService(project, nodeConfig);
   const dynamicDsService = new DynamicDsService(dsProcessorService, project);
 
   const poiService = new PoiService(nodeConfig, sequilize);
   const storeService = new StoreService(sequilize, nodeConfig);
-  const subqueryRepo = SubqueryFactory(sequilize);
   const mmrService = new MmrService(nodeConfig, sequilize);
   const sandboxService = new SandboxService(
     apiService,
@@ -152,7 +152,6 @@ function createIndexerManager(project: SubqueryProject): IndexerManager {
     storeService,
     nodeConfig,
     dynamicDsService,
-    subqueryRepo,
     eventEmitter,
   );
 
@@ -166,7 +165,6 @@ function createIndexerManager(project: SubqueryProject): IndexerManager {
     sandboxService,
     dsProcessorService,
     dynamicDsService,
-    subqueryRepo,
     projectService,
   );
 }
@@ -182,14 +180,14 @@ describe('IndexerManager', () => {
   });
 
   xit('should be able to start the manager (v0.0.1)', async () => {
-    indexerManager = createIndexerManager(testSubqueryProject_1());
+    indexerManager = createIndexerManager(testSubqueryProject_1(), nodeConfig);
     await expect(indexerManager.start()).resolves.toBe(undefined);
 
     expect(Object.keys((indexerManager as any).vms).length).toBe(1);
   });
 
   xit('should be able to start the manager (v0.2.0)', async () => {
-    indexerManager = createIndexerManager(testSubqueryProject_2());
+    indexerManager = createIndexerManager(testSubqueryProject_2(), nodeConfig);
     await expect(indexerManager.start()).resolves.toBe(undefined);
 
     expect(Object.keys((indexerManager as any).vms).length).toBe(1);