+++ packages/node/src/indexer/indexer.manager.ts
@@ -1,7 +1,7 @@
 // Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { Inject, Injectable } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { ApiPromise } from '@polkadot/api';
 import { RuntimeVersion } from '@polkadot/types/interfaces';
 import { hexToU8a, u8aEq } from '@polkadot/util';
@@ -21,9 +21,7 @@ import {
   PoiBlock,
   StoreService,
   PoiService,
-  SubqueryRepo,
   NodeConfig,
-  getYargsOption,
   getLogger,
   profiler,
   profilerWrap,
@@ -40,6 +38,7 @@ import {
   SubqueryProject,
 } from '../configure/SubqueryProject';
 import * as SubstrateUtil from '../utils/substrate';
+import { yargsOptions } from '../yargs';
 import { ApiService } from './api.service';
 import {
   asSecondLayerHandlerProcessor_1_0_0,
@@ -53,7 +52,6 @@ import { ApiAt, BlockContent } from './types';
 const NULL_MERKEL_ROOT = hexToU8a('0x00');
 
 const logger = getLogger('indexer');
-const { argv } = getYargsOption();
 
 @Injectable()
 export class IndexerManager {
@@ -71,14 +69,14 @@ export class IndexerManager {
     private sandboxService: SandboxService,
     private dsProcessorService: DsProcessorService,
     private dynamicDsService: DynamicDsService,
-    @Inject('Subquery') protected subqueryRepo: SubqueryRepo,
     private projectService: ProjectService,
   ) {
     logger.info('indexer manager start');
+
     this.api = this.apiService.getApi();
   }
 
-  @profiler(argv.profiler)
+  @profiler(yargsOptions.argv.profiler)
   async indexBlock(
     blockContent: BlockContent,
     runtimeVersion: RuntimeVersion,
@@ -299,7 +297,7 @@ export class IndexerManager {
 
       for (const handler of handlers) {
         vm = vm ?? (await getVM(ds));
-        argv.profiler
+        this.nodeConfig.profiler
           ? await profilerWrap(
               vm.securedExec.bind(vm),
               'handlerPerformance',
@@ -352,6 +350,7 @@ export class IndexerManager {
     ) => boolean,
   ): SubstrateCustomHandler[] {
     const plugin = this.dsProcessorService.getDsProcessor(ds);
+
     return ds.mapping.handlers
       .filter((handler) => {
         const processor = plugin.handlerProcessors[handler.kind];
@@ -365,6 +364,7 @@ export class IndexerManager {
         const processor = asSecondLayerHandlerProcessor_1_0_0(
           plugin.handlerProcessors[handler.kind],
         );
+
         try {
           return processor.filterProcessor({
             filter: handler.filter,
@@ -386,6 +386,7 @@ export class IndexerManager {
   ): Promise<void> {
     const plugin = this.dsProcessorService.getDsProcessor(ds);
     const assets = await this.dsProcessorService.getAssets(ds);
+
     const processor = asSecondLayerHandlerProcessor_1_0_0(
       plugin.handlerProcessors[handler.kind],
     );