+++ packages/node/src/indexer/sandbox.service.ts
@@ -7,13 +7,7 @@ import {
   isDatasourceV0_2_0,
   SubstrateDataSource,
 } from '@subql/common-substrate';
-import {
-  timeout,
-  NodeConfig,
-  StoreService,
-  getYargsOption,
-  getLogger,
-} from '@subql/node-core';
+import { timeout, NodeConfig, StoreService, getLogger } from '@subql/node-core';
 import { Store } from '@subql/types';
 import { levelFilter } from '@subql/utils';
 import { merge } from 'lodash';
@@ -23,8 +17,6 @@ import { getProjectEntry } from '../utils/project';
 import { ApiService } from './api.service';
 import { ApiAt } from './types';
 
-const { argv } = getYargsOption();
-
 export interface SandboxOption {
   store?: Store;
   script: string;
@@ -32,27 +24,33 @@ export interface SandboxOption {
   entry: string;
 }
 
-const DEFAULT_OPTION: NodeVMOptions = {
-  console: 'redirect',
-  wasm: argv.unsafe,
-  sandbox: {},
-  require: {
-    builtin: argv.unsafe
-      ? ['*']
-      : ['assert', 'buffer', 'crypto', 'util', 'path'],
-    external: true,
-    context: 'sandbox',
-  },
-  wrapper: 'commonjs',
-  sourceExtensions: ['js', 'cjs'],
+const DEFAULT_OPTION = (nodeConfig: NodeConfig): NodeVMOptions => {
+  return {
+    console: 'redirect',
+    wasm: nodeConfig?.unsafe,
+    sandbox: {},
+    require: {
+      builtin: nodeConfig.unsafe
+        ? ['*']
+        : ['assert', 'buffer', 'crypto', 'util', 'path'],
+      external: true,
+      context: 'sandbox',
+    },
+    wrapper: 'commonjs',
+    sourceExtensions: ['js', 'cjs'],
+  };
 };
 
 const logger = getLogger('sandbox');
 
 export class Sandbox extends NodeVM {
-  constructor(option: SandboxOption, protected readonly script: VMScript) {
+  constructor(
+    option: SandboxOption,
+    protected readonly script: VMScript,
+    protected config: NodeConfig,
+  ) {
     super(
-      merge(DEFAULT_OPTION, {
+      merge(DEFAULT_OPTION(config), {
         require: {
           root: option.root,
           resolve: (moduleName: string) => {
@@ -69,7 +67,7 @@ export class Sandbox extends NodeVM {
 }
 
 export class IndexerSandbox extends Sandbox {
-  constructor(option: SandboxOption, private readonly config: NodeConfig) {
+  constructor(option: SandboxOption, config: NodeConfig) {
     super(
       option,
       new VMScript(
@@ -78,6 +76,7 @@ export class IndexerSandbox extends Sandbox {
     `,
         path.join(option.root, 'sandbox'),
       ),
+      config,
     );
     this.injectGlobals(option);
   }
@@ -135,7 +134,7 @@ export class SandboxService {
       this.processorCache[entry] = processor;
     }
     processor.freeze(api, 'api');
-    if (argv.unsafe) {
+    if (this.nodeConfig.unsafe) {
       processor.freeze(this.apiService.getApi(), 'unsafeApi');
     }
     return processor;