+++ packages/node/src/indexer/worker/worker.module.ts
@@ -10,13 +10,7 @@ import { IndexerModule } from '../indexer.module';
 
 @Module({
   imports: [
-    DbModule.forRoot({
-      host: process.env.DB_HOST ?? '127.0.0.1',
-      port: process.env.DB_PORT ? Number(process.env.DB_PORT) : 5432,
-      username: process.env.DB_USER ?? 'postgres',
-      password: process.env.DB_PASS ?? 'postgres',
-      database: process.env.DB_DATABASE ?? 'postgres',
-    }),
+    DbModule.forRoot(),
     EventEmitterModule.forRoot(),
     ConfigureModule.register(),
     ScheduleModule.forRoot(),