+++ test/jest-setup.ts
@@ -1,2 +1,5 @@
 import 'reflect-metadata';
-import 'regenerator-runtime/runtime'
+import 'regenerator-runtime/runtime';
+
+// Causes the logger to be init
+(global as any).__TEST__ = true;