+++ packages/cli/CHANGELOG.md
@@ -8,6 +8,18 @@ All logs must start with the format: [x.y.z] - yyyy-mm-dd
 
 ## [Unreleased]
 
+## [1.6.4] - 2022-10-06
+- Update IPFS endpoints. (#1337)
+
+## [1.6.3] - 2022-09-29
+### Fixed
+- Fix Output URL for `project:create-project` command. (#1324)
+- Improve `codegen` command error logs. (#1324)
+
+## [1.6.2] - 2022-09-24
+### Fixed
+- Fixed cli publish failed due to upload file logic unable to validate section is `null`. (#1316)
+
 ## [1.6.1] - 2022-09-13
 ### Fixed
 - Fixed `subql-init` for Algorand, Avalanche and Cosmos.