+++ packages/query/CHANGELOG.md
@@ -8,6 +8,12 @@ All logs must start with the format: [x.y.z] - yyyy-mm-dd
 
 ## [Unreleased]
 
+## [1.6.0]
+### Changed
+- Change max connection to 10 ms and query timeout to 10000 ms. (#1322)
+### Added
+- Added hot-schema reload, query-service no longer needs to restart to see schema changes. (#1321)
+
 ## [1.5.0] - 2022-08-11
 ### Fixed
 - Fix db selection in configuration when subscription is enabled (#1241)