+++ packages/utils/CHANGELOG.md
@@ -8,6 +8,10 @@ All logs must start with the format: [x.y.z] - yyyy-mm-dd
 
 ## [Unreleased]
 
+## [1.2.0] - 2022-08-11
+### Changed
+- Update `sequelize` to 6.23.0. (#1311)
+
 ## [1.1.0] - 2022-08-11
 ### Changed
 - Update `@polkadot/util` to v10 (#1230)